requirements.txt
flask
flask-cors
flask-swagger-ui
groq
python-dotenv
